refactor: extract 3 utility modules from friday-agent-runtime (5248→3782 lines)
Extract ~1,466 lines of pure utility functions from the monolithic
agent runtime into focused, independently testable modules:

- friday-agent-tool-call-recovery.ts (116 lines): JSON parsing for
  recovering tool calls from plain text LLM output
- friday-agent-time-sensitive-handler.ts (422 lines): timezone handling,
  news intent classification, response timeliness validation
- friday-agent-closure-gap-detector.ts (1001 lines): evidence validation,
  closure gap detection, boundary enforcement, desktop inspection

All extracted functions are pure or read-only with zero coupling to the
core executeRun() closure state. The core loop and all public APIs remain
unchanged. 9,645 tests pass, 0 failures.
